A key vulnerability escaped me
Time is a funtion of this app.This measures the success  of the app.

Application available on the Playstore have time vulnerability
Let us say an app is based on a Trial Basis,then the functions of the app terminate as soon as the trial is approached..

A cool way to manipulate the trial is to change the System time when ever the app is in usage.This fools the internal trial function as it depends on System time

Another thing is the TimeStamp
-Either UNIX Epoch Time
-SHA-256 Generated Hash
This prevents the faking of time.